Added sign in tests
